+++ deb/deb.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/blakesmith/ar"
 	"github.com/goreleaser/nfpm"
-	"github.com/goreleaser/nfpm/pkgutil"
+	"github.com/goreleaser/nfpm/glob"
 	"github.com/pkg/errors"
 )
 
@@ -88,8 +88,8 @@ func createDataTarGz(info nfpm.Info) (dataTarGz, md5sums []byte, instSize int64,
 		info.Files,
 		info.ConfigFiles,
 	} {
-		for glob, dstroot := range files {
-			globbed, err := pkgutil.Glob(glob, dstroot)
+		for srcglob, dstroot := range files {
+			globbed, err := glob.Glob(srcglob, dstroot)
 			if err != nil {
 				return nil, nil, 0, err
 			}
