+++ glob/glob_test.go
@@ -1,4 +1,4 @@
-package pkgutil
+package glob
 
 import (
 	"testing"
